use implicit value classes for extension methods
this way there's no allocation of the extension classes at runtime.
verified with profiler and the following test app:

```
object Foo extends App {
  val gratefulDead = GratefulDead.traversal(GratefulDead.newGraphWithData)
  val artists = gratefulDead.artists
  artists.l
  println("connect now")
  Thread.sleep(20_000)
  println("starting")

  while (true) {
    artists.sangSongs.l
  }
}
```
