+++ traversals/src/main/scala/io/shiftleft/overflowdb/traversals/Traversal.scala
@@ -24,5 +24,3 @@ object Traversal extends IterableFactory[Traversal] {
   def newBuilder[A]: mutable.Builder[A, Traversal[A]] = Iterator.newBuilder[A].mapResult(new Traversal(_))
   def from[A](source: IterableOnce[A]): Traversal[A] = new Traversal(Iterator.from(source))
 }
-
-
