+++ traversals/src/main/scala/io/shiftleft/overflowdb/traversals/package.scala
@@ -1,6 +1,26 @@
 package io.shiftleft.overflowdb
 
 package object traversals {
-  implicit def traversalToNodeTraversal[A <: NodeRef[_]](traversal: Traversal[A]): NodeTraversal[A] =
-    new NodeTraversal[A](traversal)
+
+  implicit class NodeTraversal[+A <: NodeRef[_]](val traversal: Traversal[A]) extends AnyVal {
+    def id: Traversal[Long] = traversal.map(_.id)
+    def label: Traversal[String] = traversal.map(_.label)
+
+    def property[P](key: String): Traversal[P] = traversal.map(_.value[P](key))
+    def hasProperty(key: String): Traversal[A] = traversal.filter(_.property(key).isPresent)
+    def hasProperty[P](key: String, value: String): Traversal[A] =
+      traversal.filter { node =>
+        val property = node.property[P](key)
+        property.isPresent && property.value == value
+      }
+
+    /** Note: do not use as the first step in a traversal, e.g. `traversalSource.all.id(value)`.
+     *  Use `traversalSource.withId` instead, it is much faster */
+    def id(value: Long): Traversal[A] = traversal.filter(_.id == value)
+
+    /** Note: do not use as the first step in a traversal, e.g. `traversalSource.all.label(value)`.
+     *  Use `traversalSource.withLabel` instead, it is much faster */
+    def label(value: String): Traversal[A] = traversal.filter(_.label == value)
+  }
+
 }
